updated the wasteService Methods

+++ src/main/java/com/enviro/assessment/grad001/MosaMoime/Enviro/service/WasteService.java
@@ -24,20 +24,33 @@ public class WasteService {
      }
 
 
-    public Iterable<WasteCategory> getWasteServiceCategories(){
-        return wasteRespostory.findAll();
+    public Iterable<WasteCategory> getWasteServiceCategories(String categoryType){
+        List<WasteCategory> filteredCategories = new ArrayList<>();
+        for (WasteCategory wasteCategory : wasteRespostory.findAll()) {
+            if (wasteCategory.getCategoryType().equalsIgnoreCase(categoryType)) {
+                filteredCategories.add(wasteCategory);
+            }
+        }
+        return filteredCategories;
     }
 
-    public WasteCategory getWasteServiceCategory(Integer id){
-        return wasteRespostory.findById(id).orElse(null);
+    public WasteCategory getWasteServiceCategory(Integer id, String categoryType){
+        WasteCategory wasteCategory = wasteRespostory.findById(id).orElse(null);
+        assert wasteCategory != null;
+        if (wasteCategory.getCategoryType().equalsIgnoreCase(categoryType) && wasteCategory.getId() == id) {
+            return wasteCategory;
+        }
+        return null;
     }
 
     public void addWasteServiceCategory(WasteCategory wasteCategory){
         wasteRespostory.save(wasteCategory);
     }
 
-    public void deleteWasteServiceCategory(Integer id){
-        wasteRespostory.deleteById(id);
+    public void deleteWasteServiceCategory(Integer id, String categoryType){
+        if (wasteRespostory.findById(id).orElse(null).getCategoryType().equalsIgnoreCase(categoryType)) {
+            wasteRespostory.deleteById(id);
+        }
     }
 
     public void updateWasteServiceCategory(Integer id, WasteCategory wasteCategory){
